绩效考核代码优化
绩效考核代码优化

+++ KPI-ejb/src/java/cn/hanbell/kpi/entity/ScorecardDetail.java
@@ -84,7 +84,7 @@ public class ScorecardDetail extends SuperDetailEntity {
     @Basic(optional = false)
     @NotNull
     @Column(name = "weight")
-    private int weight;
+    private BigDecimal weight;
     @Column(name = "categoryId")
     private Integer categoryId;
     @Size(max = 45)
@@ -405,11 +405,11 @@ public class ScorecardDetail extends SuperDetailEntity {
         this.kind = kind;
     }
 
-    public int getWeight() {
+    public BigDecimal getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(BigDecimal weight) {
         this.weight = weight;
     }
